chore: also honor existing documents for other types
