+++ modules/ingestor/src/graph/mod.rs
@@ -9,6 +9,7 @@ pub mod vulnerability;
 
 use sea_orm::DbErr;
 use std::fmt::Debug;
+use std::ops::{Deref, DerefMut};
 
 #[derive(Debug, Clone)]
 pub struct Graph {
@@ -26,5 +27,40 @@ pub enum Error<E: Send> {
 impl Graph {
     pub fn new(db: trustify_common::db::Database) -> Self {
         Self { db }
+    }
+}
+
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub enum Outcome<T> {
+    Existed(T),
+    Added(T),
+}
+
+impl<T> Outcome<T> {
+    pub fn into_inner(self) -> T {
+        match self {
+            Outcome::Existed(value) => value,
+            Outcome::Added(value) => value,
+        }
+    }
+}
+
+impl<T> Deref for Outcome<T> {
+    type Target = T;
+
+    fn deref(&self) -> &Self::Target {
+        match self {
+            Outcome::Existed(value) => value,
+            Outcome::Added(value) => value,
+        }
+    }
+}
+
+impl<T> DerefMut for Outcome<T> {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        match self {
+            Outcome::Existed(value) => value,
+            Outcome::Added(value) => value,
+        }
     }
 }
