+++ modules/ingestor/src/service/sbom/clearly_defined.rs
@@ -1,8 +1,5 @@
 use crate::{
-    graph::{
-        sbom::{Outcome, SbomInformation},
-        Graph,
-    },
+    graph::{sbom::SbomInformation, Graph, Outcome},
     model::IngestResult,
     service::Error,
 };
